changed all projection matrix to have infinity far
A shadow artifacts that the shadow go away to the point at infinity would be found when I see the shadow from caster to reciver. because I didn't use a same the projection matrix at both drawing shadow volume and receiver objects, so I have made it same. (Normal projection matrix made a z component to be greater than 1.0 so it will be clipped in NDC space)